Reorganized actions files
