+++ src/actions/mod.rs
@@ -1,9 +1,12 @@
 use bevy::prelude::*;
 
-use crate::actions::game_control::{get_movement, GameControl};
 use crate::GameState;
 
-mod game_control;
+pub mod resources;
+mod systems;
+
+use resources::*;
+use systems::*;
 
 pub struct ActionsPlugin;
 
@@ -14,32 +17,5 @@ impl Plugin for ActionsPlugin {
         app.init_resource::<Actions>()
             .add_system(handle_escape)
             .add_system(set_movement_actions.in_set(OnUpdate(GameState::Playing)));
-    }
-}
-
-#[derive(Default, Resource)]
-pub struct Actions {
-    pub player_movement: Option<Vec2>,
-    pub exit_game: bool,
-}
-
-pub fn set_movement_actions(mut actions: ResMut<Actions>, keyboard_input: Res<Input<KeyCode>>) {
-    let player_movement = Vec2::new(
-        get_movement(GameControl::Right, &keyboard_input)
-            - get_movement(GameControl::Left, &keyboard_input),
-        get_movement(GameControl::Up, &keyboard_input)
-            - get_movement(GameControl::Down, &keyboard_input),
-    );
-
-    if player_movement != Vec2::ZERO {
-        actions.player_movement = Some(player_movement.normalize());
-    } else {
-        actions.player_movement = None;
-    }
-}
-
-pub fn handle_escape(mut actions: ResMut<Actions>, keyboard_input: Res<Input<KeyCode>>) {
-    if (GameControl::Escape.pressed(&keyboard_input)) {
-        actions.exit_game = true;
     }
 }
