+++ src/actions/resources.rs
@@ -1,0 +1,7 @@
+use bevy::prelude::*;
+
+#[derive(Default, Resource)]
+pub struct Actions {
+    pub player_movement: Option<Vec2>,
+    pub exit_game: bool,
+}
